Move files to proper locations
